+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # Release notes
 
+## 0.6.0
+add access logger method
+
+### Depricated
+`preset/healthz` use plugin/healthcheck
+`preset/status` use plugin/status
+
 ## 0.5.0
 `formBodyPipe` pipe to parse form data